fix: address review feedback on the 0.76.1 bugfix merge (#1789)
- api: require trusted-local access for GET /api/diagnostics/network so
  remote clients cannot read network and peer diagnostics under listen_all;
  add the route to the sensitive-route regression test.
- identity: propagate node-key-path resolution failures through
  identity_state_dir / was_public_path and their callers instead of
  silently falling back to the default ~/.mesh-llm namespace.
  load_last_mesh_id logs and returns None (no sticky mesh) on failure.
- identity crate: treat the empty ancestor of a relative key path as the
  existing current directory so `keys/node.key` creates ./keys instead of
  failing with NotFound; add a regression test.
- doctor: keep configured_selection as the raw config value and report the
  --llama-flavor override as a separate effective_selection field.
- doctor: move the network doctor implementation out of crates/mesh-llm
  into mesh_llm_commands::doctor; the binary crate keeps dispatch only.
- scripts: remove the primary library from the ordered list by path
  identity; pin readelf to LC_ALL=C in both ELF inspectors.
- tests: pick the foreign architecture relative to the host, and assert
  the glibc control case exits 0.
- regenerate the console print allowlist for the moved lines.
