+++ crates/mesh-llm-commands/src/runtime_native.rs
@@ -389,7 +389,8 @@ pub fn run_native_runtime_doctor(
         running_mesh_version: CURRENT_MESH_VERSION.to_string(),
         selected_mesh_version: selected_mesh_version.to_string(),
         configured_skippy_abi: skippy_abi_version.map(ToString::to_string),
-        configured_selection: effective_selection.map(ToString::to_string),
+        configured_selection: configured_selection.map(ToString::to_string),
+        effective_selection: effective_selection.map(ToString::to_string),
         host: profile,
         cache_path: cache.root().to_path_buf(),
         selected_runtime_id: selected.map(|runtime| runtime.native_runtime_id.clone()),
