+++ tools/xtask/data/console_print_allowlist.json
@@ -661,6 +661,16 @@
       "macro_name": "println!"
     }
   ],
+  "crates/mesh-llm-commands/src/doctor.rs": [
+    {
+      "line": 10,
+      "macro_name": "println!"
+    },
+    {
+      "line": 14,
+      "macro_name": "println!"
+    }
+  ],
   "crates/mesh-llm-commands/src/gpus.rs": [
     {
       "line": 29,
@@ -1163,23 +1173,19 @@
   ],
   "crates/mesh-llm-commands/src/runtime_native/formatters.rs": [
     {
-      "line": 90,
+      "line": 94,
       "macro_name": "eprintln!"
     },
     {
-      "line": 92,
+      "line": 96,
       "macro_name": "eprintln!"
     },
     {
-      "line": 113,
+      "line": 117,
       "macro_name": "eprintln!"
     },
     {
-      "line": 114,
-      "macro_name": "eprintln!"
-    },
-    {
-      "line": 121,
+      "line": 118,
       "macro_name": "eprintln!"
     },
     {
@@ -1187,39 +1193,35 @@
       "macro_name": "eprintln!"
     },
     {
-      "line": 127,
+      "line": 129,
       "macro_name": "eprintln!"
     },
     {
-      "line": 138,
+      "line": 131,
       "macro_name": "eprintln!"
     },
     {
-      "line": 140,
+      "line": 142,
       "macro_name": "eprintln!"
     },
     {
-      "line": 149,
+      "line": 144,
       "macro_name": "eprintln!"
     },
     {
-      "line": 151,
+      "line": 153,
       "macro_name": "eprintln!"
     },
     {
-      "line": 156,
+      "line": 155,
       "macro_name": "eprintln!"
     },
     {
-      "line": 235,
-      "macro_name": "println!"
+      "line": 160,
+      "macro_name": "eprintln!"
     },
     {
-      "line": 248,
-      "macro_name": "println!"
-    },
-    {
-      "line": 249,
+      "line": 239,
       "macro_name": "println!"
     },
     {
@@ -1227,27 +1229,27 @@
       "macro_name": "println!"
     },
     {
-      "line": 260,
+      "line": 253,
       "macro_name": "println!"
     },
     {
-      "line": 265,
+      "line": 256,
       "macro_name": "println!"
     },
     {
-      "line": 270,
+      "line": 264,
       "macro_name": "println!"
     },
     {
-      "line": 273,
+      "line": 269,
       "macro_name": "println!"
     },
     {
-      "line": 280,
+      "line": 274,
       "macro_name": "println!"
     },
     {
-      "line": 281,
+      "line": 277,
       "macro_name": "println!"
     },
     {
@@ -1259,7 +1261,11 @@
       "macro_name": "println!"
     },
     {
-      "line": 287,
+      "line": 288,
+      "macro_name": "println!"
+    },
+    {
+      "line": 289,
       "macro_name": "println!"
     },
     {
@@ -1267,19 +1273,15 @@
       "macro_name": "println!"
     },
     {
-      "line": 298,
-      "macro_name": "eprintln!"
+      "line": 295,
+      "macro_name": "println!"
     },
     {
       "line": 302,
       "macro_name": "eprintln!"
     },
     {
-      "line": 303,
-      "macro_name": "eprintln!"
-    },
-    {
-      "line": 304,
+      "line": 306,
       "macro_name": "eprintln!"
     },
     {
@@ -1291,11 +1293,15 @@
       "macro_name": "eprintln!"
     },
     {
-      "line": 309,
+      "line": 311,
       "macro_name": "eprintln!"
     },
     {
-      "line": 310,
+      "line": 312,
+      "macro_name": "eprintln!"
+    },
+    {
+      "line": 313,
       "macro_name": "eprintln!"
     },
     {
@@ -1303,20 +1309,8 @@
       "macro_name": "eprintln!"
     },
     {
-      "line": 319,
-      "macro_name": "println!"
-    },
-    {
-      "line": 320,
-      "macro_name": "println!"
-    },
-    {
-      "line": 321,
-      "macro_name": "println!"
-    },
-    {
-      "line": 322,
-      "macro_name": "println!"
+      "line": 318,
+      "macro_name": "eprintln!"
     },
     {
       "line": 323,
@@ -1327,23 +1321,35 @@
       "macro_name": "println!"
     },
     {
-      "line": 329,
+      "line": 325,
       "macro_name": "println!"
     },
     {
-      "line": 332,
+      "line": 326,
       "macro_name": "println!"
     },
     {
-      "line": 335,
+      "line": 327,
       "macro_name": "println!"
     },
     {
-      "line": 337,
+      "line": 328,
       "macro_name": "println!"
     },
     {
-      "line": 338,
+      "line": 333,
+      "macro_name": "println!"
+    },
+    {
+      "line": 336,
+      "macro_name": "println!"
+    },
+    {
+      "line": 339,
+      "macro_name": "println!"
+    },
+    {
+      "line": 344,
       "macro_name": "println!"
     },
     {
@@ -1351,15 +1357,11 @@
       "macro_name": "println!"
     },
     {
-      "line": 349,
+      "line": 347,
       "macro_name": "println!"
     },
     {
-      "line": 351,
-      "macro_name": "println!"
-    },
-    {
-      "line": 354,
+      "line": 355,
       "macro_name": "println!"
     },
     {
@@ -1367,11 +1369,11 @@
       "macro_name": "println!"
     },
     {
-      "line": 361,
+      "line": 360,
       "macro_name": "println!"
     },
     {
-      "line": 362,
+      "line": 363,
       "macro_name": "println!"
     },
     {
@@ -1379,23 +1381,35 @@
       "macro_name": "println!"
     },
     {
-      "line": 368,
-      "macro_name": "println!"
-    },
-    {
       "line": 370,
       "macro_name": "println!"
     },
     {
-      "line": 374,
+      "line": 371,
       "macro_name": "println!"
     },
     {
-      "line": 375,
+      "line": 376,
       "macro_name": "println!"
     },
     {
       "line": 377,
+      "macro_name": "println!"
+    },
+    {
+      "line": 379,
+      "macro_name": "println!"
+    },
+    {
+      "line": 383,
+      "macro_name": "println!"
+    },
+    {
+      "line": 384,
+      "macro_name": "println!"
+    },
+    {
+      "line": 386,
       "macro_name": "println!"
     }
   ],
@@ -2173,31 +2187,23 @@
   ],
   "crates/mesh-llm/src/commands/doctor.rs": [
     {
-      "line": 71,
+      "line": 82,
       "macro_name": "println!"
     },
     {
-      "line": 75,
+      "line": 85,
       "macro_name": "println!"
     },
     {
-      "line": 194,
+      "line": 88,
       "macro_name": "println!"
     },
     {
-      "line": 197,
+      "line": 89,
       "macro_name": "println!"
     },
     {
-      "line": 200,
-      "macro_name": "println!"
-    },
-    {
-      "line": 201,
-      "macro_name": "println!"
-    },
-    {
-      "line": 203,
+      "line": 91,
       "macro_name": "println!"
     }
   ],
